requirements.txt
albumentations==1.4.18
einops==0.8.1
jsmin==3.0.1
matplotlib==3.10.6
ml_collections==1.1.0
networkx==3.5
numpy==2.3.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.2
Pillow==11.3.0
pycocotools==2.0.10
scipy==1.16.1
Shapely==1.7.1
skimage==0.0
thop==0.1.1.post2209072238
timm==1.0.19
torch==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.2
